Separate release tag service from public metadata release tag service.
