+++ spec/services/catalog/marc856_generator_spec.rb
@@ -145,7 +145,7 @@ RSpec.describe Catalog::Marc856Generator do
   end
 
   before do
-    allow(ReleaseTagService).to receive(:released_to_searchworks?).and_return(release_data)
+    allow(PublicMetadataReleaseTagService).to receive(:released_to_searchworks?).and_return(release_data)
   end
 
   describe '.create' do
